- export...() is now under ImageRef, as it mutates the image - bridge is no longer
