--- vips/bridge.h
@@ -1,4 +1,0 @@
-#include <stdlib.h>
-#include <vips/vips.h>
-
-void gobject_set_property(VipsObject* object, const char* name, const GValue* value);
